requirements.txt
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Computer vision
opencv-python>=4.5.0
numpy>=1.21.0
Pillow>=8.3.0

# SAM model
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# YOLO integration
ultralytics>=8.0.0

# LLaVA/Ollama integration
ollama>=0.1.0
requests>=2.28.0

# Data validation
pydantic>=2.0.0
PyYAML>=6.0

# Scientific computing
scipy>=1.7.0

# Process management
psutil>=5.9.0
